se modifico el cuit de double a float

+++ db/migrate/20160820092935_add_columns_to_users.rb
@@ -5,7 +5,7 @@ class AddColumnsToUsers < ActiveRecord::Migration[5.0]
     add_column :users, :localidad_id, :int
     add_column :users, :cuig, :string
     add_column :users, :renspa, :string
-    add_column :users, :cuit, :double
+    add_column :users, :cuit, :float
     add_column :users, :telefono, :string
     add_column :users, :codigoPostal, :string
     add_column :users, :provincia_id, :int
